Move module package and rename to meta.Package
This commit moves the module type to a a new pacakge named meta.
This package contains loose abstractions of the SPDX package to
capture their metadata.
